libbitstream: allow fields to be int or double
Change the type of the following fields to double and allow them
to appear in the JSON metadata as either int or double:
afu_image.clock_frequency_high
afu_image.clock_frequency_low
afu_image.power

+++ libbitstream/metadatav1.h
@@ -53,9 +53,9 @@ typedef struct _opae_metadata_accelerator_cluster_v1 {
  * and Partial Reconfiguration interface.
  */
 typedef struct _opae_metadata_afu_image_v1 {
-	int clock_frequency_high;
-	int clock_frequency_low;
-	int power;
+	double clock_frequency_high;
+	double clock_frequency_low;
+	double power;
 	char *interface_uuid;
 	int magic_no;
 	int num_clusters;
